Restructure LineValue
