+++ tremor-script/src/lib.rs
@@ -50,6 +50,7 @@ pub mod interpreter;
 /// The Tremor Script Lexer
 pub mod lexer;
 // We need this because of lalrpop
+mod event_payload;
 #[allow(unused)]
 pub(crate) mod parser;
 /// Support for module paths
@@ -69,6 +70,8 @@ mod tilde;
 /// Utility functions
 pub mod utils;
 
+pub use event_payload::{EventPayload, EventPayloadDeserError, ValueAndMeta};
+
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
@@ -77,7 +80,6 @@ extern crate rental;
 
 pub use crate::ast::query::SelectType;
 pub use crate::ctx::{EventContext, EventOriginUri};
-use crate::prelude::*;
 pub use crate::query::{Query, QueryRental};
 pub use crate::registry::{
     aggr as aggr_registry, registry, Aggr as AggrRegistry, CustomFn, Registry, TremorAggrFn,
@@ -88,7 +90,6 @@ use ast::{Consts, InvokeAggrFn};
 pub use interpreter::{AggrType, FALSE, NULL, TRUE};
 use lazy_static::lazy_static;
 use std::sync::atomic::{AtomicU32, Ordering};
-use std::{mem, pin::Pin};
 pub use tremor_common::stry;
 pub use tremor_value::{KnownKey, Object, Value};
 
@@ -100,221 +101,6 @@ pub static RECURSION_LIMIT: AtomicU32 = AtomicU32::new(1024);
 pub fn recursion_limit() -> u32 {
     RECURSION_LIMIT.load(Ordering::Relaxed)
 }
-
-/// Combined struct for an event value and metadata
-#[derive(
-    Clone, Debug, PartialEq, Serialize, simd_json_derive::Serialize, simd_json_derive::Deserialize,
-)]
-pub struct ValueAndMeta<'event> {
-    v: Value<'event>,
-    m: Value<'event>,
-}
-
-impl simd_json_derive::Serialize for LineValue {
-    fn json_write<W>(&self, writer: &mut W) -> std::io::Result<()>
-    where
-        W: std::io::Write,
-    {
-        self.rent(|d| d.json_write(writer))
-    }
-}
-
-impl<'input> simd_json_derive::Deserialize<'input> for LineValue {
-    fn from_tape(tape: &mut simd_json_derive::Tape<'input>) -> simd_json::Result<Self>
-    where
-        Self: Sized + 'input,
-    {
-        let ValueAndMeta { v, m } = simd_json_derive::Deserialize::from_tape(tape)?;
-
-        Ok(Self::new(vec![], |_| {
-            ValueAndMeta::from_parts(v.into_static(), m.into_static())
-        }))
-    }
-}
-
-impl<'event> ValueAndMeta<'event> {
-    /// A value from it's parts
-    #[must_use]
-    pub fn from_parts(v: Value<'event>, m: Value<'event>) -> Self {
-        Self { v, m }
-    }
-    /// Event value
-    #[must_use]
-    pub fn value(&self) -> &Value<'event> {
-        &self.v
-    }
-
-    /// Event value
-    #[must_use]
-    pub fn value_mut(&mut self) -> &mut Value<'event> {
-        &mut self.v
-    }
-    /// Event metadata
-    #[must_use]
-    pub fn meta(&self) -> &Value<'event> {
-        &self.m
-    }
-    /// Deconstruicts the value into it's parts
-    #[must_use]
-    pub fn into_parts(self) -> (Value<'event>, Value<'event>) {
-        (self.v, self.m)
-    }
-    /// borrows both parts as mutalbe
-    #[must_use]
-    pub fn parts_mut(&mut self) -> (&mut Value<'event>, &mut Value<'event>) {
-        (&mut self.v, &mut self.m)
-    }
-}
-
-impl<'event> Default for ValueAndMeta<'event> {
-    fn default() -> Self {
-        ValueAndMeta {
-            v: Value::object(),
-            m: Value::object(),
-        }
-    }
-}
-
-impl<'v> From<Value<'v>> for ValueAndMeta<'v> {
-    fn from(v: Value<'v>) -> ValueAndMeta<'v> {
-        ValueAndMeta {
-            v,
-            m: Value::object(),
-        }
-    }
-}
-
-rental! {
-    /// Tremor script rentals to work around lifetime
-    /// issues
-    pub(crate) mod rentals {
-        use super::*;
-        use std::pin::Pin;
-
-        /// Rental wrapped value with the data it was parsed
-        /// from from
-        #[rental_mut(covariant, debug)]
-        pub struct Value {
-            raw: Vec<Pin<Vec<u8>>>,
-            parsed: ValueAndMeta<'raw>
-        }
-
-    }
-}
-
-impl rentals::Value {
-    /// Borrow the parts (event and metadata) from a rental.
-    #[must_use]
-    pub fn parts<'value, 'borrow>(&'borrow self) -> (&'borrow Value<'value>, &'borrow Value<'value>)
-    where
-        'borrow: 'value,
-    {
-        let ValueAndMeta { v, m } = self.suffix();
-        (v, m)
-    }
-
-    /// Consumes an event into another
-    /// This function works around a rental limitation that is meant
-    /// to protect its users: Rental does not allow you to get both
-    /// the owned and borrowed part at the same time.
-    ///
-    /// The reason for that is that once those are taken out of the
-    /// rental the link of lifetimes between them would be broken
-    /// and you'd risk invalid pointers or memory leaks.
-    /// We however do not really want to take them out, all we want
-    /// is combine two rentals. We can do this since:
-    /// 1) the owned values are inside a vector, while the vector
-    ///    itself may be relocated by adding to it, the values
-    ///    in it will stay in the same location.
-    /// 2) we are only ever adding / extending never deleting
-    ///    or modifying.
-    ///
-    /// So what this function does it is crowbars the content
-    /// from a rental into an accessible struct then uses this
-    /// to modify its content by adding the owned parts of
-    /// `other` into the owned part `self` and the running
-    /// a merge function on the borrowed parts
-    ///
-    /// # Errors
-    /// if `join_f` errors
-    pub fn consume<'run, E, F>(&'run mut self, other: Self, join_f: F) -> Result<(), E>
-    where
-        E: std::error::Error,
-        F: Fn(&mut ValueAndMeta<'static>, ValueAndMeta<'static>) -> Result<(), E>,
-    {
-        struct ScrewRental {
-            pub parsed: ValueAndMeta<'static>,
-            pub raw: Vec<Pin<Vec<u8>>>,
-        }
-        // ALLOW: https://github.com/tremor-rs/tremor-runtime/issues/1031
-        #[allow(clippy::transmute_ptr_to_ptr)]
-        unsafe {
-            // ALLOW: https://github.com/tremor-rs/tremor-runtime/issues/1031
-            let self_unrent: &'run mut ScrewRental = mem::transmute(self);
-            // ALLOW: https://github.com/tremor-rs/tremor-runtime/issues/1031
-            let mut other_unrent: ScrewRental = mem::transmute(other);
-            self_unrent.raw.append(&mut other_unrent.raw);
-            join_f(&mut self_unrent.parsed, other_unrent.parsed)?;
-        }
-        Ok(())
-    }
-}
-
-impl From<Value<'static>> for rentals::Value {
-    fn from(v: Value<'static>) -> Self {
-        Self::new(vec![], |_| ValueAndMeta::from(v))
-    }
-}
-
-impl<T1, T2> From<(T1, T2)> for rentals::Value
-where
-    Value<'static>: From<T1> + From<T2>,
-{
-    fn from((v, m): (T1, T2)) -> Self {
-        Self::new(vec![], |_| ValueAndMeta {
-            v: Value::from(v),
-            m: Value::from(m),
-        })
-    }
-}
-
-impl Clone for LineValue {
-    fn clone(&self) -> Self {
-        // The only safe way to clone a line value is to clone
-        // the data and then turn it into a owned value
-        // then turn this owned value back into a borrowed value
-        // we need to do this dance to 'free' value from the
-        // linked lifetime.
-        // An alternative would be keeping the raw data in an ARC
-        // instead of a Box.
-        Self::new(vec![], |_| {
-            let v = self.suffix();
-            ValueAndMeta::from_parts(v.value().clone_static(), v.meta().clone_static())
-        })
-    }
-}
-impl Default for LineValue {
-    fn default() -> Self {
-        Self::new(vec![], |_| ValueAndMeta::default())
-    }
-}
-
-/// An error occurred while deserializing
-/// a value into an Event.
-pub enum LineValueDeserError {
-    /// The value was missing the `value` key
-    ValueMissing,
-    /// The value was missing the `metadata` key
-    MetaMissing,
-}
-
-impl PartialEq for LineValue {
-    fn eq(&self, other: &Self) -> bool {
-        self.rent(|s| other.rent(|o| s == o))
-    }
-}
-
-pub use rentals::Value as LineValue;
 
 pub(crate) const NO_AGGRS: [InvokeAggrFn<'static>; 0] = [];
 lazy_static! {
@@ -328,6 +114,7 @@ mod tests {
     use crate::interpreter::AggrType;
     use crate::lexer::TokenSpan;
     use crate::path::ModulePath;
+    use crate::prelude::*;
     use halfbrown::hashmap;
 
     macro_rules! eval {
